Refactor ws
