+++ worker/Cargo.toml
@@ -13,11 +13,9 @@ worker = { version = "0.3.0", features = ["http", "axum"] }
 getrandom = { version = "0.2", features = ["js"] }
 axum = { version = "0.7", features = [
     "json",
-    "macros",
 ], default_features = false }
 tower-service = "0.3"
 http = "1.1"
-cfg-if = "1.0.0"
 futures = "0.3"
 regex = "1.10"
 wasm-bindgen-futures = "0.4"
